+++ CHANGES.md
@@ -5,6 +5,35 @@ for a full listing of issues for each release.
 
 --------------------------------------------------------------------------------
 
+## 2015.10.1 (2015-10-15)
+
+Rose release 31. This release works best with
+[cylc-6.7.1](https://github.com/cylc/cylc/releases/tag/6.7.0) and
+[fcm-2015.10.0](https://github.com/metomi/fcm/releases/tag/2015.10.0).
+
+### Noteworthy Changes
+
+[#1720](https://github.com/metomi/rose/pull/1720):
+rose macro: fix parsing of blank value repeats in settings of a `namelist:`
+section.
+
+[#1718](https://github.com/metomi/rose/pull/1718):
+rose suite-run: fix hang up of `svn diff`. Call to `svn diff` may hang if user
+has an alternate `diff` command configured that hangs. This change should force
+`svn diff` to use its internal diff logic, and should call the relevant `svn`
+commands with the `--non-interactive` options.
+
+[#1714](https://github.com/metomi/rose/pull/1714):
+rose app-run, rose suite-run, rose_arch, etc: This change allows any hash
+object in Python's `hashlib` to be used to calculate the check sum of a file.
+(Only `md5` was allowed before.)
+
+[#1728](https://github.com/metomi/rose/pull/1728),
+[#1713](https://github.com/metomi/rose/pull/1713):
+rosie copy/create: support copy to another repository.
+
+--------------------------------------------------------------------------------
+
 ## 2015.10.0 (2015-10-07)
 
 Rose release 30. This release works best with